+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/PostgresLedgerExtenderService.cs
@@ -267,6 +267,10 @@ UPDATE pending_transactions
         {
             var sw = Stopwatch.StartNew();
 
+            referencedEntities.MarkSeenAddress((EntityAddress)networkConfiguration.WellKnownAddresses.TransactionTracker);
+            referencedEntities.MarkSeenAddress((EntityAddress)networkConfiguration.WellKnownAddresses.ConsensusManager);
+            referencedEntities.MarkSeenAddress((EntityAddress)networkConfiguration.WellKnownAddresses.Xrd);
+
             foreach (var committedTransaction in ledgerExtension.CommittedTransactions)
             {
                 var stateVersion = committedTransaction.ResultantStateIdentifiers.StateVersion;
@@ -414,6 +418,14 @@ UPDATE pending_transactions
                             referencedEntity.PostResolveConfigure((GlobalFungibleResourceEntity e) =>
                             {
                                 e.Divisibility = fungibleResourceManager.Value.Divisibility;
+                            });
+                        }
+
+                        if (substateData is CoreModel.NonFungibleResourceManagerFieldMutableFieldsSubstate mutableFields)
+                        {
+                            referencedEntity.PostResolveConfigure((GlobalNonFungibleResourceEntity e) =>
+                            {
+                                e.NonFungibleDataMutableFields = mutableFields.Value.MutableFields.Select(x => x.Name).ToList();
                             });
                         }
 
@@ -599,7 +611,6 @@ UPDATE pending_transactions
                     GlobalAccountLockerEntity => CoreModel.EntityType.GlobalAccountLocker,
                     GlobalIdentityEntity => CoreModel.EntityType.GlobalIdentity,
                     GlobalAccessControllerEntity => CoreModel.EntityType.GlobalAccessController,
-
                     // skipped GlobalVirtualSecp256k1Account, GlobalVirtualSecp256k1Identity, GlobalVirtualEd25519Account and GlobalVirtualEd25519Identity as they are virtual
                     GlobalFungibleResourceEntity => CoreModel.EntityType.GlobalFungibleResource,
                     InternalFungibleVaultEntity => CoreModel.EntityType.InternalFungibleVault,
@@ -746,6 +757,8 @@ UPDATE pending_transactions
         var validatorProcessor = new ValidatorProcessor(processorContext, referencedEntities);
         var validatorEmissionProcessor = new ValidatorEmissionProcessor(processorContext);
         var accountLockerProcessor = new AccountLockerProcessor(processorContext, referencedEntities);
+        var globalEventEmitterProcessor = new GlobalEventEmitterProcessor(processorContext, referencedEntities, networkConfiguration);
+        var affectedGlobalEntitiesProcessor = new AffectedGlobalEntitiesProcessor(processorContext, referencedEntities, networkConfiguration);
 
         // step: scan all substates & events to figure out changes
         {
@@ -757,7 +770,6 @@ UPDATE pending_transactions
                 var stateUpdates = committedTransaction.Receipt.StateUpdates;
                 var events = committedTransaction.Receipt.Events ?? new List<CoreModel.Event>();
                 long? passingEpoch = null;
-                var affectedGlobalEntities = new HashSet<long>();
 
                 try
                 {
@@ -766,7 +778,6 @@ UPDATE pending_transactions
                         var substateId = substate.SubstateId;
                         var substateData = substate.Value.SubstateData;
                         var referencedEntity = referencedEntities.Get((EntityAddress)substateId.EntityAddress);
-                        affectedGlobalEntities.Add(referencedEntity.AffectedGlobalEntityId);
 
                         if (substateData is CoreModel.FungibleVaultFieldBalanceSubstate fungibleVaultFieldBalanceSubstate)
                         {
@@ -907,6 +918,7 @@ UPDATE pending_transactions
                         keyValueStoreProcessor.VisitUpsert(substateData, referencedEntity, stateVersion);
                         validatorProcessor.VisitUpsert(substateData, referencedEntity, stateVersion, passingEpoch);
                         accountLockerProcessor.VisitUpsert(substateData, referencedEntity, stateVersion);
+                        affectedGlobalEntitiesProcessor.VisitUpsert(referencedEntity, stateVersion);
                         twoWayLinkProcessor.VisitUpsert(substateData, referencedEntity, stateVersion);
                     }
 
@@ -914,7 +926,6 @@ UPDATE pending_transactions
                     {
                         var substateId = deletedSubstate.SubstateId;
                         var referencedEntity = referencedEntities.GetOrAdd((EntityAddress)substateId.EntityAddress, ea => new ReferencedEntity(ea, substateId.EntityType, stateVersion));
-                        affectedGlobalEntities.Add(referencedEntity.AffectedGlobalEntityId);
 
                         if (substateId.SubstateType == CoreModel.SubstateType.NonFungibleVaultContentsIndexEntry)
                         {
@@ -924,13 +935,14 @@ UPDATE pending_transactions
                             vaultSnapshots.Add(new NonFungibleVaultSnapshot(referencedEntity, resourceEntity, simpleRep, true, stateVersion));
                         }
 
+                        affectedGlobalEntitiesProcessor.VisitDelete(referencedEntity, stateVersion);
                         packageCodeProcessor.VisitDelete(substateId, referencedEntity, stateVersion);
                         entitySchemaProcessor.VisitDelete(substateId, referencedEntity, stateVersion);
                     }
 
                     var transaction = ledgerTransactionsToAdd.Single(x => x.StateVersion == stateVersion);
 
-                    transaction.AffectedGlobalEntities = affectedGlobalEntities.ToArray();
+                    transaction.AffectedGlobalEntities = affectedGlobalEntitiesProcessor.GetAllAffectedGlobalEntities(stateVersion).ToArray();
                     transaction.ReceiptEventEmitters = events.Select(e => e.Type.Emitter.ToJson()).ToArray();
                     transaction.ReceiptEventNames = events.Select(e => e.Type.Name).ToArray();
                     transaction.ReceiptEventSbors = events.Select(e => e.Data.GetDataBytes()).ToArray();
@@ -939,17 +951,10 @@ UPDATE pending_transactions
                     transaction.ReceiptEventTypeIndexes = events.Select(e => e.Type.TypeReference.FullTypeId.LocalTypeId.Id).ToArray();
                     transaction.ReceiptEventSborTypeKinds = events.Select(e => e.Type.TypeReference.FullTypeId.LocalTypeId.Kind.ToModel()).ToArray();
 
-                    ledgerTransactionMarkersToAdd.AddRange(
-                        affectedGlobalEntities.Select(
-                            affectedEntity => new AffectedGlobalEntityTransactionMarker
-                    {
-                        Id = sequences.LedgerTransactionMarkerSequence++,
-                        EntityId = affectedEntity,
-                        StateVersion = stateVersion,
-                    }));
-
                     foreach (var @event in events)
                     {
+                        globalEventEmitterProcessor.VisitEvent(@event, stateVersion);
+
                         if (@event.Type.Emitter is not CoreModel.MethodEventEmitterIdentifier methodEventEmitter
                             || methodEventEmitter.ObjectModuleId != CoreModel.ModuleId.Main
                             || methodEventEmitter.Entity.EntityType == CoreModel.EntityType.GlobalGenericComponent
@@ -1175,6 +1180,8 @@ UPDATE pending_transactions
             validatorProcessor.ProcessChanges();
             validatorEmissionProcessor.ProcessChanges();
             accountLockerProcessor.ProcessChanges();
+            ledgerTransactionMarkersToAdd.AddRange(globalEventEmitterProcessor.CreateTransactionMarkers());
+            ledgerTransactionMarkersToAdd.AddRange(affectedGlobalEntitiesProcessor.CreateTransactionMarkers());
             twoWayLinkProcessor.ProcessChanges();
 
             foreach (var e in nonFungibleIdChanges)