+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Migrations/20240722083807_InitialCreate.cs
@@ -91,7 +91,7 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                 .Annotation("Npgsql:Enum:ledger_transaction_marker_event_type", "withdrawal,deposit")
                 .Annotation("Npgsql:Enum:ledger_transaction_marker_operation_type", "resource_in_use,account_deposited_into,account_withdrawn_from,account_owner_method_call,badge_presented")
                 .Annotation("Npgsql:Enum:ledger_transaction_marker_origin_type", "user,epoch_change")
-                .Annotation("Npgsql:Enum:ledger_transaction_marker_type", "origin,event,manifest_address,affected_global_entity,manifest_class")
+                .Annotation("Npgsql:Enum:ledger_transaction_marker_type", "origin,event,manifest_address,affected_global_entity,manifest_class,event_global_emitter")
                 .Annotation("Npgsql:Enum:ledger_transaction_status", "succeeded,failed")
                 .Annotation("Npgsql:Enum:ledger_transaction_type", "genesis,user,round_update,flash")
                 .Annotation("Npgsql:Enum:module_id", "main,metadata,royalty,role_assignment")
@@ -283,7 +283,8 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                     blueprint_version = table.Column<string>(type: "text", nullable: true),
                     assigned_module_ids = table.Column<List<ModuleId>>(type: "module_id[]", nullable: true),
                     divisibility = table.Column<int>(type: "integer", nullable: true),
-                    non_fungible_id_type = table.Column<NonFungibleIdType>(type: "non_fungible_id_type", nullable: true)
+                    non_fungible_id_type = table.Column<NonFungibleIdType>(type: "non_fungible_id_type", nullable: true),
+                    non_fungible_data_mutable_fields = table.Column<List<string>>(type: "text[]", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -1105,6 +1106,12 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
 
             migrationBuilder.CreateIndex(
                 name: "IX_ledger_transaction_markers_entity_id_state_version",
+                table: "ledger_transaction_markers",
+                columns: new[] { "entity_id", "state_version" },
+                filter: "discriminator = 'event_global_emitter'");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_ledger_transaction_markers_entity_id_state_version1",
                 table: "ledger_transaction_markers",
                 columns: new[] { "entity_id", "state_version" },
                 filter: "discriminator = 'affected_global_entity'");