+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Models/KeyValueStoreEntryDefinition.cs
@@ -64,7 +64,6 @@
 
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Diagnostics.CodeAnalysis;
 
 namespace RadixDlt.NetworkGateway.PostgresIntegration.Models;