+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Services/StandardMetadataResolver.cs
@@ -1,3 +1,67 @@
+/* Copyright 2021 Radix Publishing Ltd incorporated in Jersey (Channel Islands).
+ *
+ * Licensed under the Radix License, Version 1.0 (the "License"); you may not use this
+ * file except in compliance with the License. You may obtain a copy of the License at:
+ *
+ * radixfoundation.org/licenses/LICENSE-v1
+ *
+ * The Licensor hereby grants permission for the Canonical version of the Work to be
+ * published, distributed and used under or by reference to the Licensor’s trademark
+ * Radix ® and use of any unregistered trade names, logos or get-up.
+ *
+ * The Licensor provides the Work (and each Contributor provides its Contributions) on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied,
+ * including, without limitation, any warranties or conditions of TITLE, NON-INFRINGEMENT,
+ * MERCHANTABILITY, or FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * Whilst the Work is capable of being deployed, used and adopted (instantiated) to create
+ * a distributed ledger it is your responsibility to test and validate the code, together
+ * with all logic and performance of that code under all foreseeable scenarios.
+ *
+ * The Licensor does not make or purport to make and hereby excludes liability for all
+ * and any representation, warranty or undertaking in any form whatsoever, whether express
+ * or implied, to any entity or person, including any representation, warranty or
+ * undertaking, as to the functionality security use, value or other characteristics of
+ * any distributed ledger nor in respect the functioning or value of any tokens which may
+ * be created stored or transferred using the Work. The Licensor does not warrant that the
+ * Work or any use of the Work complies with any law or regulation in any territory where
+ * it may be implemented or used or that it will be appropriate for any specific purpose.
+ *
+ * Neither the licensor nor any current or former employees, officers, directors, partners,
+ * trustees, representatives, agents, advisors, contractors, or volunteers of the Licensor
+ * shall be liable for any direct or indirect, special, incidental, consequential or other
+ * losses of any kind, in tort, contract or otherwise (including but not limited to loss
+ * of revenue, income or profits, or loss of use or data, or loss of reputation, or loss
+ * of any economic or other opportunity of whatsoever nature or howsoever arising), arising
+ * out of or in connection with (without limitation of any use, misuse, of any ledger system
+ * or use made or its functionality or any performance or operation of any code or protocol
+ * caused by bugs or programming or logic errors or otherwise);
+ *
+ * A. any offer, purchase, holding, use, sale, exchange or transmission of any
+ * cryptographic keys, tokens or assets created, exchanged, stored or arising from any
+ * interaction with the Work;
+ *
+ * B. any failure in a transmission or loss of any token or assets keys or other digital
+ * artefacts due to errors in transmission;
+ *
+ * C. bugs, hacks, logic errors or faults in the Work or any communication;
+ *
+ * D. system software or apparatus including but not limited to losses caused by errors
+ * in holding or transmitting tokens by any third-party;
+ *
+ * E. breaches or failure of security including hacker attacks, loss or disclosure of
+ * password, loss of private key, unauthorised use or misuse of such passwords or keys;
+ *
+ * F. any losses including loss of anticipated savings or other benefits resulting from
+ * use of the Work or any changes to the Work (however implemented).
+ *
+ * You are solely responsible for; testing, validating and evaluation of all operation
+ * logic, functionality, security and appropriateness of using the Work for any commercial
+ * or non-commercial purpose and for any reproduction or redistribution by You of the
+ * Work. You assume all risks associated with Your use of the Work and the exercise of
+ * permissions under this License.
+ */
+
 using Newtonsoft.Json;
 using RadixDlt.NetworkGateway.Abstractions;
 using RadixDlt.NetworkGateway.Abstractions.StandardMetadata;
@@ -15,7 +79,7 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Services;
 
 internal class StandardMetadataResolver
 {
-    private record PartiallyValidatedStandardMetadata(long FromStateVersion, StandardMetadataKey Discriminator, bool IsLocked, string EntityAddress, string TargetValue, string ValidationResult);
+    private record PartiallyValidatedTwoWayLink(long FromStateVersion, StandardMetadataKey Discriminator, bool IsLocked, string EntityAddress, string TargetValue, string ValidationResult);
 
     private readonly ReadOnlyDbContext _dbContext;
     private readonly IDapperWrapper _dapperWrapper;
@@ -26,7 +90,7 @@ internal class StandardMetadataResolver
         _dapperWrapper = dapperWrapper;
     }
 
-    public async Task<IDictionary<EntityAddress, ICollection<ResolvedTwoWayLink>>> Resolve(
+    public async Task<IDictionary<EntityAddress, ICollection<ResolvedTwoWayLink>>> ResolveTwoWayLinks(
         ICollection<Entity> entities,
         bool resolveValidOnly,
         bool validateOnLedgerOnly,
@@ -41,7 +105,7 @@ internal class StandardMetadataResolver
         }
 
 #pragma warning disable SA1118
-        var partiallyValidatedEntries = await _dapperWrapper.ToList<PartiallyValidatedStandardMetadata>(
+        var partiallyValidatedEntries = await _dapperWrapper.ToList<PartiallyValidatedTwoWayLink>(
             _dbContext,
             @"
 WITH
@@ -234,7 +298,12 @@ WHERE @includeInvalid OR validation_result IN (@validationOnLedgerSucceeded, @va
 
         await Parallel.ForEachAsync(partiallyValidatedEntries, options, async (pv, innerToken) =>
         {
-            var resolved = await Resolve(pv, validateOnLedgerOnly, partiallyValidatedEntries, innerToken);
+            var resolved = await ResolveTwoWayLink(pv, validateOnLedgerOnly, partiallyValidatedEntries, innerToken);
+
+            if (resolved == null)
+            {
+                return;
+            }
 
             result.GetOrAdd((EntityAddress)pv.EntityAddress, _ => new ConcurrentQueue<ResolvedTwoWayLink>()).Enqueue(resolved);
         });
@@ -242,7 +311,7 @@ WHERE @includeInvalid OR validation_result IN (@validationOnLedgerSucceeded, @va
         return result.ToDictionary(e => e.Key, e => (ICollection<ResolvedTwoWayLink>)e.Value.ToList());
     }
 
-    private async ValueTask<ResolvedTwoWayLink> Resolve(PartiallyValidatedStandardMetadata entry, bool validateOnLedgerOnly, ICollection<PartiallyValidatedStandardMetadata> allEntries, CancellationToken token)
+    private async ValueTask<ResolvedTwoWayLink?> ResolveTwoWayLink(PartiallyValidatedTwoWayLink entry, bool validateOnLedgerOnly, ICollection<PartiallyValidatedTwoWayLink> allEntries, CancellationToken token)
     {
         if (entry.ValidationResult == StandardMetadataConstants.ValidationUnknown)
         {
@@ -273,7 +342,7 @@ WHERE @includeInvalid OR validation_result IN (@validationOnLedgerSucceeded, @va
 
         return entry.Discriminator switch
         {
-            StandardMetadataKey.DappAccountType => new DappAccountMarkerResolvedTwoWayLink(invalidReason),
+            StandardMetadataKey.DappAccountType => null,
             StandardMetadataKey.DappDefinition => new DappDefinitionResolvedTwoWayLink((EntityAddress)entry.TargetValue, invalidReason),
             StandardMetadataKey.DappDefinitions => new DappDefinitionsResolvedTwoWayLink((EntityAddress)entry.TargetValue, invalidReason),
             StandardMetadataKey.DappClaimedEntities => new DappClaimedEntityResolvedTwoWayLink((EntityAddress)entry.TargetValue, invalidReason),
@@ -315,14 +384,14 @@ WHERE @includeInvalid OR validation_result IN (@validationOnLedgerSucceeded, @va
         return new DappClaimedWebsiteResolvedTwoWayLink(new Uri(claimedWebsite), invalidReason);
     }
 
-    private DappAccountLockerResolvedTwoWayLink ResolveDappAccountLocker(EntityAddress entityAddress, EntityAddress lockerAddress, ICollection<PartiallyValidatedStandardMetadata> allEntries)
+    private DappAccountLockerResolvedTwoWayLink ResolveDappAccountLocker(EntityAddress entityAddress, EntityAddress lockerAddress, ICollection<PartiallyValidatedTwoWayLink> allEntries)
     {
         var valid = allEntries.Any(x => x.EntityAddress == entityAddress && x.Discriminator == StandardMetadataKey.DappClaimedEntities && x.TargetValue == lockerAddress);
 
         return new DappAccountLockerResolvedTwoWayLink(entityAddress, valid ? null : "claimed_entities entry with the locker address missing");
     }
 
-    private Exception CreateException(PartiallyValidatedStandardMetadata entry, string details)
+    private Exception CreateException(PartiallyValidatedTwoWayLink entry, string details)
     {
         var dump = JsonConvert.SerializeObject(entry);