+++ CHANGELOG.md
@@ -3,18 +3,26 @@ Release built: _not released yet_
 
 - Optimized `/statistics/validators/uptime` endpoint processing time.
 
+### Breaking Changes
+- `/stream/transactions` no longer indexes `affected_global_entities` for the transaction tracker and the consensus manager entity types.
+
 ### Bug fixes
 - Properly indexes manifest classes. Some transactions might have been previously misclassified as `Transfer` and `AccountDepositSettingsUpdate`, i.e. empty transactions with only `lock_fee` instruction.
 
 ### API Changes
 - Added support for the missing `message` and `flags.disable_auth_checks` properties in the `/transaction/preview` endpoint request.
+- Added list of mutable non fungible data fields `non_fungible_data_mutable_fields` returned from `/state/entity/details` endpoint.
+- New `event_global_emitters_filter` filter added to `/stream/transactions` endpoint. It allows filtering transactions by the global ancestor of an event emitter. For events emitted by a global entity it is going to be that entity, for internal entities it is going to be a global ancestor.
 - Added support for two-way linked dApps in the `/state/entity/details` endpoint.
   - Brand-new `two_way_linked_*` properties on the `details` property of Resources, Accounts, Packages and other global components.
   - See https://docs.radixdlt.com/docs/metadata-for-verification#metadata-standards-for-verification-of-onledger-entities for detailed specification.
 
+
 ### Database changes
 - Replaced relationship-related columns (`*_entity_id`) in the `entities` table with more generic collection implementation using `correlated_entity_*` columns.
 - Replaced per-epoch validator emissions (`validator_emission_statistics` table) with their cumulative statistics (`validator_cumulative_emission_history` table).
+- Added `non_fungible_data_mutable_fields` to `entities` table. Which contains list of all mutable non fungible data fields for non fungible resource entities.
+- New `ledger_transaction_markers` type with the `event_global_emitter` discriminator. It represents the global emitter for each event.
 
 ## 1.6.1
 Release built: 21.05.2024